requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
deep-translator>=1.11.0
pyjwt>=2.8.0
bcrypt>=4.1.0
